requirements.txt
aiodns>=3.0.0
aiofiles>=0.6.0
aiohttp>=3.7.4.post0
aiosqlite>=0.17.0
async-generator>=1.10
async-timeout>=3.0.1
asyncio-extras>=1.3.2
asyncpraw>=7.4.0
asyncprawcore>=2.3.0
attrs>=21.2.0
Brotli>=1.0.9
cchardet>=2.1.7
certifi>=2021.10.8
cffi>=1.15.0
chardet>=4.0.0
charset-normalizer>=2.0.7
dnspython>=2.1.0
ffmpeg>=1.4
greenlet>=1.1.2
html2text>=2020.1.16
idna>=3.3
Markdown>=3.3.4
msgpack>=1.0.2
multidict>=5.2.0
nextcord>=2.0.0a3
orjson>=3.6.4
psutil>=5.8.0
pycares>=4.1.2
pycparser>=2.20
pymongo>=3.12.1
PyNaCl>=1.4.0
requests>=2.26.0
ruamel.yaml>=0.17.17
six>=1.16.0
typing-extensions>=3.10.0.2
update-checker>=0.18.0
urllib3>=1.26.7
vacefron.py>=1.6.2
websocket-client>=1.2.1
yarl>=1.7.2
youtube-dl>=2021.6.6
python-dateutil>=2.8.2
jaraco.context>=4.1.1
more-itertools>=8.10.0
websockets>=10.0
wolframalpha>=5.0.0
xmltodict>=0.12.0